Add topbar for admin dashboard and loading state UI
The topbar needs to be modified since the admin dashboard is a protected
route. The logged out user should not be able to access the admin
dashboard after logging out, so the logout should redirect to the home
page.

The loading state UI is added to the admin dashboard to get rid of the
flickering effect when the user authentication state is being checked so
users don't see the unauthorized page for a split second even though
it is still getting the user and not because they don't have permission.