Fix bug with signing in and out creating tabs
The ideal tab behavior when signing out would clear all tabs back to
the default single tab. But the code doesn't differentiate between
different users since we don't store user ID in local storage such
that when we log out the tab numbering would continue from
the previously logged in user.

Now we track user ID with local storage to fix the issue and check
for tabs in the editor when logging out so that we don't create new
tabs if there are tabs already in the editor. Still need to reset tabs
though.