Fix mark solve and unmark solve issue in UI
We have a lot of duplicate problem state from using the get problem
data hook, we only refetch from the top level hook's data (workspace)
so the hook data in the child components don't get updated and thus
the UI doesn't get updated for those components. To fix all the dupe
state from the child components need to get raised all the way to the
workspace level and then use refetch on that data and pass the state
down to the children. The children delete their own dupe state
and rely on workspace's version of state which is the only one
that exists and gets refreshed for proper rendering