Fix topbar arrow navigation to different problems
Previous implementation relied on consecutive order ids for it to find
the next or previous problem (next problem has to be current order
plus 1). This isn't necessarily true all the time. The problems in the
problem sets are organized in sorted ascending order, so the next
problem is the next order number in the current sorted order.
Changed implementation to account for this.

Also should probably build on it so that problem navigation only
navigates to problems in the current problem set rather than all
problems as it is currently doing, but will require keeping track of
current problem set, filtering the problem order ids only in that
problem set, and then using that for nav.