Add side nav bar for problems in problem workspace
Reused some code from the homepage. Had to refactor some hooks to use
in the side nav bar. Cleaned up some files as well.

Overflow still not working in the side nav bar when expanding accordions
and the button in the top bar to open the side nav bar has sizing issues
since it should just fit the height of the parent topbar and the image
shouldn't be popping out and overflowing so will have to fix. Also
the background for the modal doesn't cover entire screen so that is
probably why other things on screen are still interactable as well.