Fix bug where accordion render early
Sometimes the accordion of problems would render early and the number of
solved problems doesn't render correctly since the hook would still be
running in the background. Tried to reuse the problem loading state but
this means that if one loading hook finishes faster than the other, it
would set loading to false even if the other is still running and
render incomplete data.

Solution is to just make a separate loading state and add loading to the
get solved problems hook.