Add trie problems and scroll to current problem
When opening side nav menu it will scroll to the current problem for
better user experience. Have to ID refs with problemSetID and problemID
since we share the same problems between different sets so we end
up scrolling to the latest set's problem that has the same ID as the
current problem we have open since we overwrite the refs or something.
We have to differentiate problems between sets so that it can scroll to
the problem in the expanded accordion and not the other accordions
with the same problem.

For future optimization, some state from the problemListDropdown
and ProblemTable can probably be lifted up to ProblemSetAccordionList
like the current problem set id from the query params which we can get
once in the whole list of accordions and then pass down as props instead
of doing it in every dropdown which is redundant, but on the other hand
it means the component is self sufficient so there is that too...

Maybe consider removing highlighting for the same problem in the
side nav menu that are in different problem sets (all the sets that are
not expanded and are not the current set) but since the problem is
technically the same in those sets too it might be worth keeping
around as well to mark where the user is in that set...