Add backtracking and heap problems, optimize code
Optimized code by lifting up calculations in the ProblemList and Problem
Table components to the ProblemSetAccordionList component so that they
are only done once and then passed down as props to the children
rather than having the code in the children where it will run for every
accordion that we have (the calculation is for getting the problem id
and set for the side nav menu which doesn't change when on the
workspace and the user is trying to open the menu. They are stuck
on that screen with the same pathname/pid and problem set so just
do it once and pass it down to be rendered.